feat: extract Silent_Skin class to separate class file
