+++ packages/wp-plugin/ionos-essentials/inc/dashboard/blocks/vulnerability/index.php
@@ -2,6 +2,9 @@
 
 namespace ionos\essentials\dashboard\blocks\vulnerability;
 
+use ionos\essentials\dashboard\Silent_Skin;
+use const ionos\essentials\PLUGIN_DIR;
+
 require_once ABSPATH . 'wp-admin/includes/plugin.php';
 require_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
 require_once ABSPATH . 'wp-admin/includes/plugin-install.php';
@@ -40,32 +43,7 @@ require_once ABSPATH . 'wp-admin/includes/misc.php';
 
 function install_plugin_from_url($plugin_url)
 {
-  class Silent_Skin extends \WP_Upgrader_Skin
-  {
-    public function feedback($string, ...$args)
-    {
-    }
-
-    public function header()
-    {
-    }
-
-    public function footer()
-    {
-    }
-
-    public function error($errors)
-    {
-    }
-
-    public function before()
-    {
-    }
-
-    public function after()
-    {
-    }
-  }
+  require_once PLUGIN_DIR . '/inc/dashboard/class-silent-skin.php';
 
   $skin     = new Silent_Skin();
   $upgrader = new \Plugin_Upgrader($skin);
