+++ packages/wp-plugin/ionos-essentials/inc/dashboard/class-silent-skin.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace ionos\essentials\dashboard;
+
+class Silent_Skin extends \WP_Upgrader_Skin
+{
+  public function feedback($string, ...$args)
+  {
+  }
+
+  public function header()
+  {
+  }
+
+  public function footer()
+  {
+  }
+
+  public function error($errors)
+  {
+  }
+
+  public function before()
+  {
+  }
+
+  public function after()
+  {
+  }
+}
